pyproject.toml
[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version"
]
update_changelog_on_bump = true

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \.venv
  | __pycache__
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
minversion = "9.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
    "."
]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "migrations",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nerdlock"
version = "0.1.0"
description = "Smart door control system with Nuki integration"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi==25.1.0",
    "Flask==3.1.2",
    "Flask-Login==0.6.3",
    "Flask-Mail==0.10.0",
    "Flask-Security-Too==5.7.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-WTF==1.2.2",
    "pyNukiBT==0.0.19",
    "bleak==0.22.3",
    "PyNaCl==1.5.0",
    "python-dotenv==1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.1",
    "pytest-flask==1.3.0",
    "pre-commit>=3.5.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "commitizen>=3.32.0",
]
